pip-requirements.txt
streamlit>=1.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
requests>=2.31.0
numpy>=1.24.0
torch>=2.0.0
